requirements.txt
adal>=1.2.1
aiohttp>=3.6.2
aiounittest>=1.3.0
arrow>=0.15.6
async-timeout>=3.0.1
asyncio>=3.4.3
attrs>=19.3.0
azure-cognitiveservices-language-luis>=0.2.0
azure-cognitiveservices-speech>=1.13.0
azure-common>=1.1.25
azure-cosmos>=3.1.2
azure-storage-blob>=2.1.0
azure-storage-common>=2.1.0
binaryornot>=0.4.4
botbuilder-ai>=4.9.1
botbuilder-azure>=4.9.0
botbuilder-core>=4.8.0
botbuilder-schema>=4.8.0
botframework-connector>=4.8.0
certifi>=2020.4.5.1
cffi>=1.14.0
chardet>=3.0.4
click>=7.1.2
config>=0.5.0.post0
cookiecutter>=1.7.2
cryptography>=2.8
datamodels>=0.8.2
idna>=2.8
isodate>=0.6.0
Jinja2>=2.11.2
jinja2-time>=0.2.0
jsonpickle>=1.2
MarkupSafe>=1.1.1
msal>=1.1.0
msrest>=0.6.10
msrestazure>=0.6.3
multidict>=4.7.5
oauthlib>=3.1.0
pip>=20.2.2
poyo>=0.5.0
pycparser>=2.20
PyJWT>=1.5.3
python-dateutil>=2.8.1
python-slugify>=4.0.0
requests>=2.22.0
requests-oauthlib>=1.3.0
setuptools>=46.1.3
six>=1.14.0
text-unidecode>=1.3
urllib3>=1.25.9
wheel>=0.34.2
yarl>=1.4.2